Create palindrome_number.cpp
This is problem from the LeetCode app called Palindrome Number. In this task l was supposed to check if the integer is palindrome, the integer is palindrome when it reads the same forward and backward.At the and in the main function is going to write "is a palindrome" if the number is palindrome and "is not palindrome" if is not.