Create roman_to_integer.cpp
This is the problem from LeetCode application where you have to translate Roman numbers to integer numbers. In this task l used the method where first of all l was checking if the next character is longer than the string itself, if is not then we have to check if this character is smaller then next one, if it is we have to subtract number from the total number. And if one of this conditions isn't fulfilled we are going just to add the current number to the main one.