Create longestCommonPrefix.cpp
This is the problem from the LeetCode app. In this task you have to find longest common prefix,  first of all it is finding the minimal length of the strings, after that is is going from every letter and the word to find matchings letters, and if it does it is going to write them in the output. For this task l used c++ language.