Create validParentheses.cpp
This is the problem from LeetCode application. In this task l was supposed to find matching parentheses for the brackets, if the program  finds matching brackets it is going to return a true value and if the program do not find matching brackets it is going to return a false value.