Create mergeTwoSortedLists.cpp
I created a struct ListNode which has two variables, one is for the value and the other one is a pointer to the next value.
In this task l had two separated lists with numbers, l had to make one singly linked list that would store all the values in the in ascending order. At the end program is going to output all numbers until ListNode reach NULL value.