Create removeDuplicatesromSortedArray
This problem is from the LeetCode app. In this problem l was supposed to remove duplicates from the non-decreasing array.
I just replaced duplicated values with the unique one and used k to keep track of them, at the end l returned number of unique elements which is used later in the main function.