Create removeElement.c
This is a problem from LeetCode application. In this task you are supposed to remove all the elements that are equal to number val. After that you have to return number of elements which are not equal to val, that number is going to be used later in the main function.