Create plusOne.c
This is a problem from the LeetCode application. In this task l was supposed to increment every digit of the given array and to return incremented array at the end and the returnSize thought paramater. In the case that incremented number would take more place than given array, than l should make another array with function malloc and make the first digit 1 and every other 0.