Create climbingStairs.c
This is a problem from the LeetCode application. In this task l was supposed to find in how many ways we can climb to the top, but l was only allowed to take 1 or 2 steps at the time. I used logic that remebers the last two steps and  just collect them so you can get a final number.