Create removeDuplicatesfromSortedList.c
This problem is from the LeetCode application. In this task l was supposed to remove duplicates from the sorted array. I used algorithm where l was comparing tail and tail->next values, if they where the same l would change tail->next to tail->next->next and in this way l would ship duplicated elements. And if not l would just change current tail to the next one.