Create mergeSortedArray.c
This is a problem from the LeetCode application. In this task l was supposed to merge two sorted arrays, the first one had positive numbers and zeros which indexes are supposed to be used for the numbers in the second array. I used the logic of comparing two numbers from the arrays and placing the bigger one on the edge of the array.