Create binaryTreeInorderTraversal.c
This is a problem from the LeetCode application. In this task l was supposed to return unordered array from the stuct called TreeNode with the pointer to the first element(root of binary tree). I did that using two recursive functions, one for the number of elements with only one argument(pointer to the root), and the second one for filling the array with 3 arguments(pointer to the current root, pointer to the result array and pointer to the number of elemets(k). I also used malloc for the allocation of the memory.