Create sameTree.c
This is a problem from the LeetCode application. In this problem l was supposed to check if the all nodes are the same, including parents and children. I did that using recursive bool function that returns if the check is true or false. At the and l just set the result to the answer of this function.