Create symmetricTree.c
This is a problem from the LeetCode application. In this task l was supposed to check of the tree was symmetrical. I did that using recursive function with two pointers, one for the left child and the other one for the right child. I passed those pointers thought main function with root->left and root->right