Create convertSortedArraytoBinarySearchTree.c
This is a problem from the LeetCode application. In this task I was supposed to convert an array (which was sorted in acceding order) to a height-balanced binary search tree. I did that by using recursive function which would always allocate more space in the mamory for the whole struct and store the value if the start variable is not greater that the end variable.