Create minimumDepthofBinaryTree.c
This is a problem from the LeetCode application. In this task l was supposed to find the minimum depth of binary tree. I did that using recursive function which would return 0 if the root was NULL, and if only one of them was NULL it would use max function and if both existed it would use min function.