Create pathSum.c
This is a problem from the LeetCode application. In this task l was supposed to find if a root-to-leaf path has sum is equal to the targetSum. I did that by the help of the recursive function which would take away root->val from the remainingSum. At the end if the remainingSum was equal to 0 it would return 1(true) else it would return 0(false).